Make it easier to use
