+++ README.md
@@ -10,31 +10,16 @@ First, add the library to your project:
 
 ### Actor lookup
 
-Then, you will need a way to find actors inside your step definitions. For example, you can add a lookup map in your World instance and a custom parameter type:
+Then, you will need a way to find actors inside your step definitions. The following snippet handles the setup for you:
 
 ```typescript
-class World {
-  private readonly actorByName = new Map<string, Actor>()
+import { ActorWorld, defineActorParameterType } from '@cucumber/screenplay'
+import { setWorldConstructor, defineParameterType } from '@cucumber/cucumber'
 
-  findOrCreateActor(actorName: string): Actor<World> {
-    let actor = this.actorByName.get(actorName)
-
-    if (actor === undefined) {
-      actor = new Actor<World>(this, actorName)
-      this.actorByName.set(actorName, actor)
-    }
-
-    return actor
-  }
-}
-
-defineParameterType({
-  name: 'actor',
-  regexp: /[A-Z][a-z]+/,
-  transformer(this: World, actorName: string) {
-    return this.findOrCreateActor(actorName)
-  },
-})
+// Sets a default World that can look up actors by name
+setWorldConstructor(ActorWorld)
+// Create a default parameter type that can map any name with an actor
+defineActorParameterType(defineParameterType)
 ```
 
 In your step, you will then be able to get your `Actor` instance, for example:
@@ -60,6 +45,8 @@ Then('{actor} should be logged-in', function (actor: Actor<World>) {
   assert.ok(actor.ask(isLoggedIn())
 })
 ```
+
+For more example of integrating, see the section dedicated to integrating this library in more complex project below.
 
 ### Creating interactions or questions
 
@@ -124,3 +111,62 @@ Then('{actor} should be logged-in', function (actor: Actor<World>) {
 
 **Note:** the data remembered are scoped by `Actor`, so you can not access a data remembered by an actor when using another one. You can also have multiple
 actors storing different data with the same key.
+
+
+## Integrating in an existing project
+
+### Adding actor lookup to an existing project
+
+If you have already started using Cucumber in your project and already created a World object, you will not be able to simply use `ActorWorld` as
+the default World class.
+You have two solutions here:
+
+#### 1) use `ActorWorld` as the super class for your existing World
+
+One simple solution to get the `findOrCreateActor`method on your `World` object is to extend the `ActorWorld` provided by this library:
+
+```typescript
+import { ActorWorld } from '@cucumber/screenplay'
+import { setWorldConstructor } from '@cucumber/cucumber'
+
+class World extends ActorWorld {
+  // ...
+}
+
+setWorldConstructor(World)
+```
+
+#### 2) Add an `ActorLookup` to your World
+
+If you can not extend `ActorWorld`, you can also add an `ActorLookup` instance to your world object like so:
+
+```typescript
+import { ActorLookUp } from '@cucumber/screenplay'
+
+class World {
+  private readonly actorLookUp = new ActorLookUp()
+
+  findOrCreateActor(actorName: string): Actor<any> {
+    return this.actorLookUp.findOrCreateActor(this, actorName)
+  }
+}
+```
+
+## Specifying different regular expressions for matching actors
+
+The regular expression used to match an `Actor` name provided by `defineActorParameterType` only match a capital latin letter followed
+by any number of latin letters. This may not fit your needs, but it is pretty easy to write your own matcher.
+
+For example, if you want to only match some specific names for your actors:
+
+```typescript
+import { setWorldConstructor, defineParameterType } from '@cucumber/cucumber'
+
+defineParameterType({
+  name: 'actor',
+  regexp: /Alice|Bob|Charlie/,
+  transformer(this: World, actorName: string) {
+    return this.findOrCreateActor(actorName)
+  },
+})
+```
