+++ src/index.ts
@@ -1,4 +1,11 @@
 import Actor, { Interaction } from './Actor'
+import ActorWorld, { ActorLookup } from './ActorWorld'
+import defineActorParameterType from './registerActorParameterType'
 
 export type { Interaction }
-export { Actor }
+export {
+  Actor,
+  ActorWorld,
+  ActorLookup,
+  defineActorParameterType
+}
